Simulation Serialization & Changes
Added: Most of the simulation objects now have JAXB annotations for
XML-Serialization. You can serialize or deserialize a simulation by
using the profile serializer.
Added: The reputations of a member are now stored in the defining member
groups. The reputation object itself now stores the member whose
reputation is stored.