Steps towards simulation
Added: Vulnerabilities are now defined in the xml file for every part.
Added: Exploit and payload objects (while payloads are not loaded
dynamically yet).
Added: Basic simulation algorithm (very basic, most parts are tests)
using an action structure: A group (or a member, which isn't yet
implemented) has an interest > The interest object calculates the best
action for a certain member (which is selected by the main simulation
line) > The action gets executed by the member it was calculated for
(e.g. an exploit attack, which is the only action at this time).