Brain Data & Knowledge
Added: Interests are now stored as "brain data entries" in the member
object. Brain data is the foundation for more things a human brain can
hold.
Added: A very, very basic class for knowledge which doesn't do anything
yet. The whole new package will be filled in the next commits.