First Iteration of the Property System
Added: There's now a really fancy property system called "World". As the name suggests, the system manages classes which can have properties. Those classes generally represent objects which are located in a world. Properties are dynamic objects of different types which store data, so modifications will be able to hook in and add their own properties.