Tick Property
Added: The tick property for executing ticks on a world object. The property will be called by the tick simulator later on when we come to it.
Added: The object instance adapter is an utility for storing objects as their class' names and just create a new instance during unmarshalling.