Replaced the Simulation Storage with World Properties
Added: The simulation object no longer stores the members etc. It now stores a world which contains a root world object which does that. We will transfer that modularity to every single simulation world related class in the next time.
Fixed: Resolved some more JavaDoc issues and enhanced it a bit.