Replaced @Execution with Exceptions
Added: The @Execution annotation wasn't flexible enough, so we added the StopExecutionException and FunctionExecutionException. Function executors can throw a stop exception if they want to abort the execution (equal to ExecutionPolicy.THIS). If the stop exception contains a cause, a new function execution exception will be thrown. That exception can be determinated by the invoker.