Lockable Function Executors
Added: Function executors (their containers) are now lockable. That mechanism is independent from the more or less "global" lock on a whole function object. That means that a lock on an executor is not affected by setLocked(false) on the function object. You can also lock executors which aren't annotated with Lockable.