Syscall Infrastructure and UserSyscalls
Added: The infrastructure for syscalls. Syscalls are requests of a process to the operating systems (like: What is the content of that file?). The syscall classes provide function definition constants which can be used on processes (permission check) or on the os object (no permission check). Program executors ask their processes to execute the syscall. That way, there is a permission check when program executors access the operating system!
Added: The class UserSyscalls which provides syscalls for user operations, as well as user and group wrappers for permission-safe user and group modifications.