Added Configuration Class for Easy Configs
Added: A configuration class which is heavily inspired by UNIX configuration files. Configuration entries (rows or lines) are customized for different purposes. For example, a User class could be a ConfigurationEntry with different properties, like the name or the groups the user is a member in. Those properties are also called "columns" and can be edited in an external editor later on.