Replace NetworkManager with (empty) NetworkModule
Add: The NetworkModule class is a normal operating system module which will eventually take care of sending, receiving and handling packets, as well as establishing connections. However, this functionality won't be implemented before the new network system is up and running. The module class is currently empty.
Remove: The old NetworkManager because it is no longer required.