Overhaul the session class
Fix: Move the old SessionProgram ckass to the .world.comp.os package and rename it to just session.
Add: The session reads the session user from the user configuration file.
Add: The session checks a supplied password if there is a parent session and it is not a root session.