Breaking: Remove all old programs & session stuff
Remove: The old programs and sessions were created using the old program system which will be deprecated soon. We decided to remove all these old programs because the news system takes a totally different approach. However, some of the classes this commit removes could be used later on for the new display system.