Classmod Scheduler API
Add: The Scheduler API allows to schedule tasks later. They are executes after a set amount of ticks. The SchedulerUser interface extends TickUpdatable and adds an executor to the TICK_UPDATE function that updates the integrated scheduler, so classes that want to use such a scheduler only need to implement that interface.
Add: The FunctionCallScheduleTask is a simple task for the scheduler that just invokes a function without parameters. It should be used instead of plain tasks because functions are easily extendable.