Add basic FileCreateProgram for creating files
Add: The FileCreateProgram takes a path string and a file type and then creates a file using the two values.
Note: Such programs are intended to be only used by so called "actions" which invoke the programs in order to achieve something. These actions are used by the ai and by the graphical desktop programs.