Add CommonLocation for providing a common file path for programs
Add: The CommonLocation annotation sets the file path of program executors they can be commonly found under. The new annotation should be used on every program executor, so users of the programs do not need to search for it.
Add: The ExecutorUtils.getCommonLocation() utility method can be used as a quick access to the common location of a program executor.