Add type predicate shortcut to AbstractEventHandler
Add: An AbstractEventHandler constructor that takes an event type (class object) and sets the handler's predicate to a TypePredicate with that event type. There is no need for a type/predicate combination since a custom predicate has a requirement for a certain event type that functions as a "type predicate".