Remove the old world event system (BREAKING)
Remove: The world.event package which contained a basic event system that was based on Classmod. It is replaced with the more advanced bridge event system.
Note: This removal will cause a lot of errors since the code that used the world event system hasn't been yet patched to the bridge event system.